Remove redirect handling logic from ConnectionPool
Slightly refactor our PoolManager#urlopen method to turn integers into
Retry objects sooner and move our redirect handling tests for connection
pools to our pool manager so we don't lose test coverage.

Some tests were duplicated and thus not moved over. The rest were not
and so I moved them over and modified their logic to exercise the
PoolManager instead of HTTPConnectionPool

Closes #951
